1.2.6.21

+++ zp-core/setup/database.php
@@ -8,6 +8,9 @@
  * @Copyright 2016 by Stephen L Billard for use in {@link https://github.com/ZenPhoto20/ZenPhoto20 ZenPhoto20}
  *
  */
+$dbSoftware = db_software();
+$indexComments = version_compare($dbSoftware['version'], '5.5.0') >= 0;
+
 $database = $orphans = array();
 foreach (getDBTables() as $table) {
 	$tablecols = db_list_fields($table);
@@ -45,15 +48,23 @@ foreach (getDBTables() as $table) {
 		unset($index['Collation']);
 		unset($index['Cardinality']);
 		unset($index['Comment']);
+		if (!$indexComments) {
+			unset($index['Index_comment']);
+		}
 
 		switch ($keyname) {
 			case 'valid':
-				if ($table == 'administrators' && $index['Column_name'] === '`valid`,`user`') {
+			case 'user':
+				$keys = explode(',', $index['Column_name']);
+				sort($keys);
+				if ($table == 'administrators' && implode(',', $keys) === '`user`,`valid`') {
 					$index['Index_comment'] = 'zp20';
 				}
 				break;
 			case 'filename':
-				if ($table == 'images' && $index['Column_name'] === '`filename`,`albumid`') {
+				$keys = explode(',', $index['Column_name']);
+				sort($keys);
+				if ($table == 'images' && implode(',', $keys) === '`albumid`,`filename`') {
 					$index['Index_comment'] = 'zp20';
 				}
 				break;
@@ -63,7 +74,6 @@ foreach (getDBTables() as $table) {
 				}
 				break;
 		}
-
 		$database[$table]['keys'][$keyname] = $index;
 	}
 }
@@ -82,6 +92,7 @@ foreach ($metadataProviders as $source => $handler) {
 		$plugin = getPlugin($source . '.php');
 		require_once($plugin);
 		$exifvars = $handler::getMetadataFields();
+
 		foreach ($exifvars as $key => $item) {
 			if (!is_null($disable = getOption($key . '-disabled'))) {
 				$exifvars[$key][5] = !($disable & true);
@@ -129,7 +140,6 @@ foreach ($metadataProviders as $source => $handler) {
 }
 
 foreach ($template as $tablename => $table) {
-
 	$exists = array_key_exists($tablename, $database);
 	if (!$exists) {
 		$create = array();
@@ -138,7 +148,6 @@ foreach ($template as $tablename => $table) {
 	}
 	foreach ($table['fields'] as $key => $field) {
 		if ($key != 'id') {
-
 			$string = "ALTER TABLE " . prefix($tablename) . " %s `" . $field['Field'] . "` " . $field['Type'];
 			if ($field['Null'] === 'NO')
 				$string .= " NOT NULL";
@@ -158,7 +167,6 @@ foreach ($template as $tablename => $table) {
 			}
 			$addString = sprintf($string, 'ADD COLUMN') . $comment . ';';
 			$changeString = sprintf($string, "CHANGE `" . $field['Field'] . "`") . $comment . ';';
-
 			if ($exists) {
 				if (array_key_exists($key, $database[$tablename]['fields'])) {
 					if ($field != $database[$tablename]['fields'][$key]) {
@@ -189,7 +197,6 @@ foreach ($template as $tablename => $table) {
 		}
 	}
 
-
 	if (isset($table['keys'])) {
 		foreach ($table['keys'] as $key => $index) {
 			$string = "ALTER TABLE " . prefix($tablename) . ' ADD ';
@@ -205,9 +212,14 @@ foreach ($template as $tablename => $table) {
 			if (!empty($index['Sub_part'])) {
 				$k .=" (" . $index['Sub_part'] . ")";
 			}
-			$alterString = "$string`$key` ($k) COMMENT 'zp20';";
+			$alterString = "$string`$key` ($k)";
+			if ($indexComments) {
+				$alterString.=" COMMENT 'zp20';";
+			} else {
+				unset($index['Index_comment']);
+			}
 			if ($exists) {
-				if (array_key_exists($key, $database[$tablename]['keys'])) {
+				if (isset($database[$tablename]['keys'][$key])) {
 					if ($index != $database[$tablename]['keys'][$key]) {
 						$dropString = "ALTER TABLE " . prefix($tablename) . " DROP INDEX `" . $index['Key_name'] . "`;";
 						setupQuery($dropString);
@@ -217,8 +229,11 @@ foreach ($template as $tablename => $table) {
 					setupQuery($alterString);
 				}
 			} else {
-				$tableString = "  $u ($k) COMMENT 'zp20',";
-				$create[] = $tableString;
+				$tableString = "  $u ($k)";
+				if ($indexComments) {
+					$tableString .= "  COMMENT 'zp20'";
+				}
+				$create[] = $tableString . ',';
 			}
 			unset($database[$tablename]['keys'][$key]);
 		}
@@ -233,7 +248,7 @@ foreach ($template as $tablename => $table) {
 		if (array_key_exists('keys', $database[$tablename]) && !empty($database[$tablename]['keys'])) {
 			foreach ($database[$tablename]['keys'] as $index) {
 				$key = $index['Key_name'];
-				if ($index['Index_comment'] === 'zp20') {
+				if (isset($index['Index_comment']) && $index['Index_comment'] === 'zp20') {
 					$dropString = "ALTER TABLE " . prefix($tablename) . " DROP INDEX `" . $key . "`;";
 					setupQuery($dropString);
 				} else {
